refactor: extract `writeTypes` utility
