+++ packages/nuxi/src/commands/build.ts
@@ -2,7 +2,8 @@ import { resolve } from 'pathe'
 import consola from 'consola'
 import { importModule } from '../utils/cjs'
 
-import { defineNuxtCommand, invokeCommand } from './index'
+import { writeTypes } from '../utils/prepare'
+import { defineNuxtCommand } from './index'
 
 export default defineNuxtCommand({
   meta: {
@@ -12,13 +13,13 @@ export default defineNuxtCommand({
   },
   async invoke (args) {
     process.env.NODE_ENV = process.env.NODE_ENV || 'production'
-    const rootDir = args.rootDir || resolve(args._[0] || '.')
+    const rootDir = resolve(args._[0] || '.')
 
     const { loadNuxt, buildNuxt } = await importModule('@nuxt/kit', rootDir) as typeof import('@nuxt/kit')
 
-    const nuxt = args.nuxt || await loadNuxt({ rootDir })
+    const nuxt = await loadNuxt({ rootDir })
 
-    await invokeCommand('prepare', { nuxt, rootDir })
+    await writeTypes(nuxt)
 
     nuxt.hook('error', (err) => {
       consola.error('Nuxt Build Error:', err)
