+++ packages/nuxi/src/commands/prepare.ts
@@ -1,9 +1,6 @@
-import { promises as fsp } from 'fs'
-import { relative, resolve } from 'pathe'
-import { cyan } from 'colorette'
-import { TSReference } from '@nuxt/kit'
-import consola from 'consola'
-import { importModule, getModulePaths, getNearestPackage } from '../utils/cjs'
+import { resolve } from 'pathe'
+import { importModule } from '../utils/cjs'
+import { writeTypes } from '../utils/prepare'
 import { defineNuxtCommand } from './index'
 
 export default defineNuxtCommand({
@@ -14,53 +11,11 @@ export default defineNuxtCommand({
   },
   async invoke (args) {
     process.env.NODE_ENV = process.env.NODE_ENV || 'production'
-    const rootDir = args.rootDir || resolve(args._[0] || '.')
+    const rootDir = resolve(args._[0] || '.')
 
     const { loadNuxt } = await importModule('@nuxt/kit', rootDir) as typeof import('@nuxt/kit')
-    const nuxt = args.nuxt || await loadNuxt({ rootDir })
+    const nuxt = await loadNuxt({ rootDir })
 
-    const modulePaths = getModulePaths(nuxt.options.modulesDir)
-
-    const references: TSReference[] = [
-      ...nuxt.options.buildModules,
-      ...nuxt.options.modules,
-      ...nuxt.options._modules
-    ]
-      .filter(f => typeof f === 'string')
-      .map(id => ({ types: getNearestPackage(id, modulePaths)?.name || id }))
-
-    const declarations: string[] = []
-
-    await nuxt.callHook('builder:generateApp')
-    await nuxt.callHook('prepare:types', { references, declarations })
-
-    const declarationPath = resolve(`${rootDir}/nuxt.d.ts`)
-
-    const declaration = [
-      '// This file is auto generated by `nuxt prepare`',
-      '// Please do not manually modify this file.',
-      '',
-      ...references.map((ref) => {
-        if ('path' in ref) {
-          ref.path = relative(rootDir, ref.path)
-        }
-        return `/// <reference ${renderAttrs(ref)} />`
-      }),
-      ...declarations,
-      'export {}',
-      ''
-    ].join('\n')
-
-    await fsp.writeFile(declarationPath, declaration)
-
-    consola.success('Generated', cyan(relative(process.cwd(), declarationPath)))
+    await writeTypes(nuxt)
   }
 })
-
-function renderAttrs (obj) {
-  return Object.entries(obj).map(e => renderAttr(e[0], e[1])).join(' ')
-}
-
-function renderAttr (key, value) {
-  return value ? `${key}="${value}"` : ''
-}
